Move pipeline with default
