+++ test/core/pipeline/features/pipeline_run.feature
@@ -11,6 +11,7 @@ Feature: Pipeline running
         | that has no components |
         | that is linear |
         | that is really complex with lots of components, forks, and loops |
+        | that has a single component with a default input |
 
     Scenario Outline: Running a bad Pipeline
         Given a pipeline <kind>
